Django-Middleware-0x03/requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
django-filter>=23.5
drf-nested-routers>=0.93.4
djangorestframework-simplejwt>=5.3.1
Pillow>=10.2.0 